+++ oalg-top/src/OAlg/Homology/Complex.hs
@@ -1,41 +1,33 @@
 
 {-# LANGUAGE NoImplicitPrelude #-}
 
-{-# LANGUAGE
-    TypeFamilies
-  , TypeOperators
-  , MultiParamTypeClasses
-  , FlexibleInstances
-  , FlexibleContexts
-  , GADTs
-  , StandaloneDeriving
-  , DataKinds
-  , TupleSections
+{-# LANGUAGE TypeFamilies
+           , TypeOperators
+           , MultiParamTypeClasses
+           , FlexibleInstances
+           , FlexibleContexts
+           , GADTs
+           , StandaloneDeriving
+           , GeneralizedNewtypeDeriving
+           , DataKinds
 #-}
+
 
 -- |
 -- Module      : OAlg.Homology.Complex
--- Description : definition of simplical complexes.
+-- Description : complex of simplices over a set complex.
 -- Copyright   : (c) Erich Gut
 -- License     : BSD3
 -- Maintainer  : zerich.gut@gmail.com
 --
--- definition of simplical 'Complex'.
+-- complex of simplices over a set complex.
 module OAlg.Homology.Complex
   (
-    -- * Complex
-    Complex(..), cpxElem, complex
-  , cpxVertices, cpxSimplices, cpxGenerators
-
-    -- * Map
-  , ComplexMap(..)
   ) where
 
-import Control.Monad
-
 import Data.Typeable
-import Data.List as L ((++)) -- head,tail,zip,last,reverse,repeat,(++),span)
-import Data.Foldable (foldl)
+
+import Data.List as L (zip,(++),repeat)
 
 import OAlg.Prelude
 
@@ -43,440 +35,241 @@ import OAlg.Category.Map
 
 import OAlg.Data.Filterable
 
-import OAlg.Hom.Distributive ()
-
-import OAlg.Entity.Sequence hiding (span,isEmpty)
-
+-- import OAlg.Structure.Fibred
+-- import OAlg.Structure.Additive
+-- import OAlg.Structure.Vectorial
 import OAlg.Structure.PartiallyOrdered
+import OAlg.Structure.Multiplicative
+import OAlg.Structure.Ring
+
+-- import OAlg.Hom.Fibred
+-- import OAlg.Hom.Additive
+-- import OAlg.Hom.Vectorial
+
+import OAlg.Entity.Natural as N
+import OAlg.Entity.FinList as F
+import OAlg.Entity.Sequence.Set
+import OAlg.Entity.Sequence.Graph
+-- import OAlg.Entity.Sum
+import OAlg.Entity.Matrix
 
 import OAlg.Homology.Simplical
+import OAlg.Homology.SetComplex
+import OAlg.Homology.Chain
+
 
 --------------------------------------------------------------------------------
 -- Complex -
 
--- | complex over a vertex type @__x__@.
+-- | finite complexe of simplices over a set complex.
 --
---  __Properties__ Let @c = 'Complex' ('Graph' zxs)@ be in @'Complex' __x__@, then holds:
+-- __Properties__ Let @c = 'Complex' sc zxs@ be in @'Complex' __s__ __n__ __x__@, then
+-- holds:
 --
---  (1) @zxs@ is not empty.
+-- (1) @z0 '==' -1@ where @(z0,_) = 'head' zxs@.
 --
---  (2) @z0 '==' -1@ where @(z0,_) = 'head' zxs@.
+-- (2) For all @(z,'Set' sxs)@ in @zxs@ and @sx@ in @sxs@ holds:
 --
---  (3) For all @(z,'Set' sxs)@ in @zsx@ holds:
+--     (1) @'dimension' sx '==' z@.
 --
---     (1) @xsx@ is not empty.
+--     (2) @'verteices' sx@ is an element of @sc@.
 --
---     (2) @'dimension' sx '==' z@ for all @sx@ in @sxs@.
---
---  (4) For all @..(z,su)':'(z',sv)..@ in @zsx@ holds:
+-- (3) For all @..(z,su)':'(z',sv)..@ in @zsx@ holds:
 --
 --    (1) @z' '==' z '+' 1@.
 --
 --    (2) @'faces'' sv'@ is a subset of @su@.
-newtype Complex x = Complex (Graph Z (Set (Set x))) deriving (Show,Eq,Ord)
-
-
---------------------------------------------------------------------------------
--- Complex - Set - Entity -
-
-instance (Entity x, Ord x) => Validable (Complex x) where
-  valid (Complex (Graph zsx)) = Label "Complex" :<=>: case zsx of
-    []            -> Label "1" :<=>: SInvalid
-    ((z,sx):zsx') -> And [ Label "2" :<=>: (z == -1) :?> Params ["z0" := show z]
-                         , valid sx
-                         , vldDimension z sx
-                         , vldFaces z sx zsx'
-                         ]
-    where
-      vldDimension z sx
-        = And [ Label "3.1" :<=>: (not $ isEmpty sx) :?> Params ["sx":=show sx]
-              , Label "3.2" :<=>: let vDim b sx = b && (dimension sx == z) in
-                  (foldl vDim True sx) :?> Params ["z":=show z, "sx" := show sx]
-              ]
-
-      vldFaces _ _ [] = SValid
-      vldFaces z su ((z',sv):zsx')
-        = And [ valid sv
-              , Label "4.1'" :<=>: (z' == succ z) :?> Params ["z":=show z, "z'":=show z']
-              , vldDimension z' sv
-              , Label "4.2" :<=>: let fsv = faces' sv in
-                  (fsv <<= su) :?> Params ["fsv":=show (fsv // su)]
-              , vldFaces z' sv zsx'
-              ]
-                                
-instance (Entity x, Ord x) => Entity (Complex x)
-
---------------------------------------------------------------------------------
--- cpxSimplices -
-
--- | the simplices of the given complex.
-cpxSimplices :: Complex x -> Graph Z (Set (Set x))
-cpxSimplices (Complex g) = g
-
---------------------------------------------------------------------------------
--- cpxGenerators -
-
--- | the generators for the given complex.
-cpxGenerators :: (Entity x, Ord x) => Complex x -> Graph Z (Set (Set x))
-cpxGenerators (Complex g) = g // gphFaces g
-
---------------------------------------------------------------------------------
--- cpxElem -
-
--- | checking for being a simplex of the given complex.
-cpxElem :: (Entity x, Ord x) => Complex x -> Set x -> Bool
-cpxElem (Complex g) = isElem $ setIndex $ gphset g where
-  isElem :: (Entity x, Ord x) => ((Z,Set x) -> Maybe n) -> Set x -> Bool
-  isElem i = isJust . i . spxAdjDim
-
-cpx :: N -> Complex N
-cpx n = complex [Set [1..n]]
-
---------------------------------------------------------------------------------
--- complex -
-
--- | the induced complex given by a list of simplices.
-complex :: (Entity x, Ord x) => [Set x] -> Complex x
-complex = Complex . foldl (||) empty . amap1 simplices
-
---------------------------------------------------------------------------------
--- cpxVertices -
-
--- | the set of vertices of the given complex.
-cpxVertices :: Complex x -> Set x
-cpxVertices (Complex g) = case gphxs g of
-  _ : (0,vs) : _ -> Set $ join $ amap1 setxs $ setxs vs
-  _              -> Set []
-
---------------------------------------------------------------------------------
--- cpxProduct -
-
-cpxProduct :: (Entity x, Ord x, Entity y, Ord y) => Complex x -> Complex y -> Complex (x,y)
-cpxProduct a b = Complex $ filter (not . isEmpty) $ gphSetFilter (elig a b) gp where
-  gp = simplices $ Set [(x,y) | x <- setxs $ cpxVertices a, y <- setxs $ cpxVertices b]
-
-  map :: (Entity x, Ord x, Entity y, Ord y) => (x -> y) -> Map EntOrd x y
-  map = Map
-
-  elig :: (Entity x, Ord x, Entity y, Ord y) => Complex x -> Complex y -> Set (x,y) -> Bool
-  elig a b = (cpxElem a . amap1 (map fst)) && (cpxElem b . amap1 (map snd))
-
-
---------------------------------------------------------------------------------
--- ComplexMap -
-
--- | mapping between complexes, where the underlying map induces a map between the two
--- given simpliex sets.
 --
--- __Property__ Let @'ComplexMap' cx cy f@ be in
--- @'ComplexMap' ('Complex' __x__) ('Complex' __y__)@ then holds:
---
--- (1) for all simplices @s@ in @cx@ holds: @'amap1' f s@ is an element of @cy@.
---
--- __Note__ If @cx@ and @cy@ are 'valid' then it is sufficient to test the property on the
--- generators, given by @'cpxGenerators' cx@.
-data ComplexMap a b where
-  ComplexMap
-    :: Complex x -> Complex y
-    -> Map EntOrd x y
-    -> ComplexMap (Complex x) (Complex y) 
+-- (4) If @sx@ is in @__s__ __x__@ with @'vertices' sx@ is in @sc@, then @sx@ is in @c@.
+data Complex s n x
+  = Complex (SetComplex x) (FinList (n+3) (Z,Set (s x)))
+  deriving (Show,Eq,Ord)
 
-
---------------------------------------------------------------------------------
--- ComplexMap - Entity -
-
-instance Show (ComplexMap a b) where
-  show (ComplexMap a b (Map f))
-    = "ComplexMap (" ++ show a ++ ") (" ++ show b ++ ") ("
-                     ++ (show $ Graph [(v,f v) | v <- setxs $ cpxVertices a]) ++ ")"
-
-instance Eq (ComplexMap a b) where
-  ComplexMap a b (Map f) == ComplexMap a' b' (Map f')
-    = (a,b,[f v | v <- vs]) == (a',b',[f' v | v <- vs])
-    where vs = setxs $ cpxVertices a
-
-instance Ord (ComplexMap a b) where
-  compare (ComplexMap a b (Map f)) (ComplexMap a' b' (Map f'))
-    = compare (a,b,[f v | v <- vs a]) (a',b',[f' v | v <- vs a'])
-    where vs = setxs . cpxVertices
-
-instance Validable (ComplexMap a b) where
-  valid (ComplexMap a b f@(Map _)) = Label "ComplexMap" :<=>:
-    And [ valid a
-        , valid b
-        , Label "1" :<=>: (fa <<= sb) :?> Params ["fa":= show (fa // sb)]
+instance (Simplical s x, Entity x, Ord x) => Validable (Complex s n x) where
+  valid (Complex c ((z,sx):|zsx')) = Label "Complex" :<=>:
+    And [ valid c
+        , Label "1" :<=>: (z == -1) :?> Params ["z0":=show z]
+        , vldDimension cElem z sx
+        , vldFaces cElem z sx zsx'
         ]
     where
-      fa = setgph $ amap1 (map spxAdjDim . Map (amap1 f) . map snd) $ gphset $ cpxGenerators a
-      sb = cpxSimplices b
+      cElem = scxElem c
 
-      map :: (Entity x, Ord x, Entity y, Ord y) => (x -> y) -> Map EntOrd x y
-      map = Map
-
-instance (Typeable a, Typeable b) => Entity (ComplexMap a b)
-
---------------------------------------------------------------------------------
-
-a = complex [Set "ab",Set "bc",Set "ac"] :: Complex Char
-b = complex [Set [0,1],Set [1,2], Set [0,2]] :: Complex N
-
-c = cpxProduct a b
-
-p1 = ComplexMap c a (Map fst)
-p2 = ComplexMap c b (Map snd)
-
-  
-{-
-instance (Entity x, Entity y) => Validable (ComplexMap (Complex x) (Complex y)) where
-  valid (ComplexMap cx cy f@(OrdMap _ )) = Label "ComplexMap" :<=>:
-    And [ valid cx
-        , valid cy
-        , vldMapSet (cpxIndex cy) f (amap1 snd $ setxs $ cpxSetMax cx)
-{-        
-        , vldElg (Set $ cpxxs cy)
-                 (spxDimSets $ join $ amap1 (setxs . smap f . snd) $ setxs $ cpxSetMax cx)
--}
-        ]
-    where
-
-      smap :: OrdMap x y -> Set (Set x) -> Set (Set y)
-      smap f@(OrdMap _) = amap1 (OrdMap $ amap1 f)
-
-      setJoin :: [Set x] -> [x]
-      setJoin = join . amap1 setxs
-
-      vldElg :: (Entity y, Ord y) => Set ((Z,Set (Set y))) -> Set ((Z,Set (Set y))) -> Statement
-      vldElg cy fx = (fx `isSubSet` cy) :?> Params ["fx":=show (fx `setDifference` cy)]
-      
-      vldMapSet _ _ [] = SValid
-      vldMapSet elg f (ssx:ssxs)
-        = And [ vldMap elg f ssx
-              , vldMapSet elg f ssxs
+      vldDimension :: Simplical s x => (Set x -> Bool) -> Z -> Set (s x) -> Statement
+      vldDimension _ _ (Set []) = SValid
+      vldDimension cElem z (Set (sx:sxs))
+        = And [ Label "2.1" :<=>: (dimension sx == z) :?> Params ["sx":= show sx]
+              , Label "2.2" :<=>:  cElem (vertices sx) :?> Params ["sx":= show sx]
+              , vldDimension cElem z (Set sxs)
               ]
 
-      vldMap _ _ (Set [])         = SValid
-      vldMap elg f (Set (sx:sxs)) = case elg $ spxAdjDim $ amap1 f sx of
-        Nothing -> False :?> Params ["sx":=show sx]
-        Just _  -> vldMap elg f (Set sxs)
+      vldFaces :: Simplical s x
+        =>(Set x -> Bool) -> Z -> Set (s x) -> FinList n (Z,Set (s x)) -> Statement
+      vldFaces _ _ _ Nil = SValid
+      vldFaces cElem z su ((z',sv):|zsx)
+        = And [ vldDimension cElem z' sv
+              , Label "3.1" :<=>: (z' == succ z) :?> Params ["z":=show z, "z'":=show z']
+              , Label "3.2" :<=>: let fsv = faces' sv in
+                  (fsv <<= su) :?> Params ["fsv":= show (fsv // su)]
+              , vldFaces cElem z' sv zsx
+              ]
 
-instance (Entity x, Entity y) => Entity (ComplexMap (Complex x) (Complex y))
-
--}
-
---------------------------------------------------------------------------------
--- ComplexMap - Entity -
-{-
-
+instance (Simplical s x, Typeable s, Entity x, Ord x, Typeable n)
+  => Entity (Complex s n x)
 
 --------------------------------------------------------------------------------
 -- cpxxs -
 
-cpxxs :: Complex x -> [(Z,Set (Set x))]
-cpxxs (Complex zsx) = zsx
+cpxxs :: Complex s n x -> FinList (n+3) (Z,Set (s x))
+cpxxs (Complex _ zssx) = zssx
 
 --------------------------------------------------------------------------------
--- cpxVertices -
+-- toFinList3 -
 
-cpxVertices :: Complex x -> Set (Set x)
-cpxVertices (Complex zsx) = case tail zsx of
-  (_,sv):_ -> sv
-  _        -> setEmpty
-
---------------------------------------------------------------------------------
--- isVertex -
-
-isVertex :: Ord x => x -> Complex x -> Bool
-isVertex x c = Set [vertex x] `isSubSet` cpxVertices c
+-- | maps a infinite list to a finite list of @__n__ + 3@.
+toFinList3 :: Any n -> [x] -> FinList (n+3) x
+toFinList3 W0 (x:x':x'':_) = x:|x':|x'':|Nil
+toFinList3 (SW n) (x:xs)   = x :| toFinList3 n xs
+toFinList3 _ _             = throw $ ImplementationError "toFinList3"
 
 --------------------------------------------------------------------------------
--- cpxSet -
+-- simplicalComplex -
 
-cpxSet :: Complex x -> Set (Z,Set x)
-cpxSet (Complex zsx) = Set $ join $ amap1 (\(z,Set sx) -> amap1 (z,) sx) zsx
-
---------------------------------------------------------------------------------
--- cpxSetMax -
-
--- | the maximal simplices for each dimension which are not faces. They form a generator set
--- of the complex.
-cpxSetMax :: Ord x => Complex x -> Set (Z,Set (Set x))
-cpxSetMax (Complex zsx)
-  = Set $ filter ((setEmpty /=).snd) $ amap1 (uncurry diff) (zsx `zip` tail zsx') where
-    zsx' = amap1 (\(_,sx) -> (faces' sx)) zsx ++ repeat setEmpty
-    diff (z,sx) sx' = (z,sx `setDifference` sx')
+-- | the embedding of a complex of set-simplices.
+simplicalComplex :: Ord x => Any n -> SetComplex x -> Complex Set n x
+simplicalComplex n c@(SetComplex (Graph zssx))
+  = Complex c (toFinList3 n ([-1..] `L.zip` ssx))
+  where ssx = amap1 snd zssx L.++ L.repeat empty
 
 --------------------------------------------------------------------------------
--- cpxIndex -
+-- singularComplex -
 
-cpxIndex :: Ord x => Complex x -> (Z,Set x) -> Maybe N
-cpxIndex = setIndex . cpxSet
+-- | complex of simplices of the given set complex, containing maybe singular simplices.
+--
+-- __Note__
+-- (1) The cardinality of the simplical sets __explodes__ for example @__s__ = []@.
+--
+-- (2) The result of 'singularComplex' for @__s__ = 'Set'@ and 'simplicalComplex' are equal,
+-- but 'simplicalComplex' is faster.
+singularComplex :: (Simplical s x, Entity x, Ord x) => Any n -> SetComplex x -> Complex s n x
+singularComplex n c = Complex c (toFinList3 n ([-1..] `L.zip` ssx)) where
+  ssx = amap1 (filter (elg c)) ((amap1 snd $ gphxs $ simplices $ scxVertices c) L.++ L.repeat empty)
 
---------------------------------------------------------------------------------
--- cpxCards -
-
-cpxCards :: Complex x -> [(Z,N)]
-cpxCards (Complex zsx) = amap1 (\(z,s) -> (z,lengthN s)) zsx
-
---------------------------------------------------------------------------------
--- cpxEmpty -
-
-cpxEmpty :: Complex x
-cpxEmpty = Complex [(-1,Set [setEmpty])]
-
---------------------------------------------------------------------------------
--- cpxBorder -
-
-cpxBorder :: Complex x -> Maybe (Complex x)
-cpxBorder (Complex [_]) = Nothing -- empty complex has no border!
-cpxBorder (Complex ss)  = Just $ Complex (dropLast ss) where
-  dropLast []     = []
-  dropLast [_]    = []
-  dropLast (x:xs) = x : dropLast xs
+  elg :: (Simplical s x, Entity x, Ord x) => SetComplex x -> s x -> Bool
+  elg c = scxElem c . vertices
 
 
---------------------------------------------------------------------------------
--- (<|>) -
+cpxl :: Any n -> N -> Complex [] n N
+cpxl n m = singularComplex n (setComplex [Set [1..m]])
 
--- | the union of two complexes.
-(<|>) :: Ord x => Complex x -> Complex x -> Complex x
-Complex as <|> Complex bs = Complex $ uni as bs where
-  uni [] bs = bs
-  uni as [] = as
-  uni ((z,sx):zsx) ((_,sy):zsy) = (z,sx `setUnion` sy):uni zsx zsy
+cpxa :: Any n -> N -> Complex Asc n N
+cpxa n m = singularComplex n (setComplex [Set [1..m]])
 
+cpxs :: Any n -> N -> Complex Set n N
+cpxs n m = singularComplex n (setComplex [Set [1..m]])
 
+cpxs' :: Any n -> N -> Complex Set n N
+cpxs' n m = simplicalComplex n (setComplex [Set [1..m]])
 
 --------------------------------------------------------------------------------
--- cpxTerminal -
+-- cpxsBoundaryRep -
 
-cpxTerminal :: Ord x => x -> Complex x
-cpxTerminal x = complex [Set [x]]
+-- | representing boundary operator according to the given comlex.
+--
+-- __Note__ From the 'Complex'-property 3.2 follwos, that if the given complex is 'valid', then its
+-- boundary representations given by 'cpxBoundaryRep' is 'valid'
+cpxsBoundaryRep :: (Ring r, Commutative r, Simplical s x, Typeable s)
+  => Complex s n x
+  -> FinList (n+2) (Representable r (ChainHom r s) (Chain r s x) (Chain r s x))
+cpxsBoundaryRep c = repBoundary (amap1 snd $ cpxxs c) where
+  repBoundary :: (Ring r, Commutative r, Simplical s x, Typeable s)
+    => FinList (n+1) (Set (s x))
+    -> FinList n (Representable r (ChainHom r s) (Chain r s x) (Chain r s x))
+  repBoundary (_:|Nil)       = Nil
+  repBoundary (sx:|sx':|sxs) = chainHomRep (Boundary sx' sx) :| repBoundary (sx':|sxs)
 
---------------------------------------------------------------------------------
--- cpxVertexSet -
-
-cpxVertexSet :: Complex x -> Set x
-cpxVertexSet = Set . amap1 (head . toList) . setxs . cpxVertices
-
---------------------------------------------------------------------------------
--- cpxProduct -
-
-
-cpxProduct :: (Ord x, Ord y) => Complex x -> Complex y -> Complex (x,y)
-cpxProduct a b
-  = Complex
-  $ filter ((/= setEmpty) . snd)
-  $ spxFilter (feasable . snd)
-  $ simplices
-  $ Set [(x,y) | x <- setxs $ cpxVertexSet a, y <- setxs $ cpxVertexSet b]
-  where
-    ia = setIndex $ cpxSet a
-    ib = setIndex $ cpxSet b
-
-    feasable xys = (spxAdjDim xs `elem` ia) && (spxAdjDim ys `elem` ib) where
-      xs = amap1 (OrdMap fst) xys
-      ys = amap1 (OrdMap snd) xys
-
-    elem x i = case i x of
-      Nothing -> False
-      Just _  -> True
+cpxsBoundaryRepZ :: (r ~ Z, Simplical s x, Typeable s)
+  => Complex s n x
+  -> FinList (n+2) (Representable r (ChainHom r s) (Chain r s x) (Chain r s x))
+cpxsBoundaryRepZ = cpxsBoundaryRep
 
 --------------------------------------------------------------------------------
 -- ComplexMap -
 
--- | mapping between complexes, where the underlying map induces a map between the two
--- given simpliex sets.
+-- | mapping between complex of simplices, where the underlying map induces a mapping between the two
+-- given simplex sets.
 --
--- __Properties__ Let @'ComplexMap' cx cy f@ be in
--- @'ComplexMap' ('Complex' __x__) ('Complex' __y__)@ then for all simplices @s@ in @cx@ holds: 
--- @'amap1' f s@ is an element of @cy@.
+-- __Property__ Let @'ComplexMap' csx csy f@ be in
+-- @'ComplexMap' __s__ __n__ (Complex __s__ __n__ __x__) (Complex __s__ __n__ __y__)@ for a
+-- 'SimplicalTransformable __s__ __x__ __y__@ structure,then holds for all simplices @s@ in @csx@:
+-- @'amap1' f s@ is an element of @csy@.
 --
--- __Note__ If @cx@ and @cy@ are 'valid' then it is sufficient to thest the property on the
--- maximal simplexes @s@, given by @'cpxSetMax' cx@.
-data ComplexMap a b where
+-- __Note__ Because of the 'SimplicalTransformable'-property and the 'Complex'-property 4 it is
+-- sufficient that the induced set-complex map - given by 'cpxSetComplexMap' - is 'valid'. 
+data ComplexMap s n x y where
   ComplexMap
-    :: Complex x -> Complex y
-    -> OrdMap x y
-    -> ComplexMap (Complex x) (Complex y) 
+    :: Complex s n x -> Complex s n y
+    -> Map EntOrd x y
+    -> ComplexMap s n (Complex s n x) (Complex s n y)
+
+-- deriving instance (Show x, Show y) => Show (ComplexMap s n (Complex s n x) (Complex s n y))
+
+--------------------------------------------------------------------------------
+-- cpxSetComplexMap -
+
+cpxSetComplexMap :: ComplexMap s n (Complex s n x) (Complex s n y)
+  -> SetComplexMap (SetComplex x) (SetComplex y)
+cpxSetComplexMap (ComplexMap (Complex cx _) (Complex cy _) f)
+  = SetComplexMap cx cy f
 
 --------------------------------------------------------------------------------
 -- ComplexMap - Entity -
 
-instance (Show x, Show y) => Show (ComplexMap (Complex x) (Complex y)) where
-  show (ComplexMap a b f) = "ComplexMap (" ++ show a ++ ") (" ++ show b ++ ") (VertexMap "
-                          ++ show [(v,amap1 f v) | v <- setxs $ cpxVertices a] ++ ")"
-
-instance Eq (ComplexMap (Complex x) (Complex y)) where
-  ComplexMap a b f@(OrdMap _) == ComplexMap a' b' f'
-    = (a,b,[amap1 f v | v <- vs]) == (a',b',[amap1 f' v | v <- vs])
-    where vs = setxs $ cpxVertices a
-
-instance Ord (ComplexMap (Complex x) (Complex y)) where
-  compare (ComplexMap a b f@(OrdMap _)) (ComplexMap a' b' f')
-    = compare (a,b,[amap1 f $ v | v <- vs a]) (a',b',[amap1 f' v | v <- vs a'])
-    where vs = setxs . cpxVertices
-
-instance (Entity x, Entity y) => Validable (ComplexMap (Complex x) (Complex y)) where
-  valid (ComplexMap cx cy f@(OrdMap _ )) = Label "ComplexMap" :<=>:
-    And [ valid cx
-        , valid cy
-        , vldMapSet (cpxIndex cy) f (amap1 snd $ setxs $ cpxSetMax cx)
-{-        
-        , vldElg (Set $ cpxxs cy)
-                 (spxDimSets $ join $ amap1 (setxs . smap f . snd) $ setxs $ cpxSetMax cx)
--}
+instance SimplicalTransformable s x y
+  => Validable (ComplexMap s n (Complex s n x) (Complex s n y)) where
+  valid f@(ComplexMap csx csy (Map _)) = Label "ComplexMap" :<=>:
+    And [ valid csx
+        , valid csy
+        , valid $ cpxSetComplexMap f
         ]
-    where
+{-
+n :: Any N4
+n = attest
 
-      smap :: OrdMap x y -> Set (Set x) -> Set (Set y)
-      smap f@(OrdMap _) = amap1 (OrdMap $ amap1 f)
+a = complex [Set "ad"]
+b = complex [Set "abc",Set "bcd",Set "ad"]
 
-      setJoin :: [Set x] -> [x]
-      setJoin = join . amap1 setxs
-
-      vldElg :: (Entity y, Ord y) => Set ((Z,Set (Set y))) -> Set ((Z,Set (Set y))) -> Statement
-      vldElg cy fx = (fx `isSubSet` cy) :?> Params ["fx":=show (fx `setDifference` cy)]
-      
-      vldMapSet _ _ [] = SValid
-      vldMapSet elg f (ssx:ssxs)
-        = And [ vldMap elg f ssx
-              , vldMapSet elg f ssxs
-              ]
-
-      vldMap _ _ (Set [])         = SValid
-      vldMap elg f (Set (sx:sxs)) = case elg $ spxAdjDim $ amap1 f sx of
-        Nothing -> False :?> Params ["sx":=show sx]
-        Just _  -> vldMap elg f (Set sxs)
-
-instance (Entity x, Entity y) => Entity (ComplexMap (Complex x) (Complex y))
-  
---------------------------------------------------------------------------------
--- cpxMapTerminal -
-
-cpxMapTerminal :: Ord x
-  => Complex x -> ComplexMap (Complex x) (Complex ())
-cpxMapTerminal c = ComplexMap c (cpxTerminal ()) (OrdMap $ const ())
-
---------------------------------------------------------------------------------
--- cpxMapVertex -
-
-cpxMapVertex :: Ord x
-  => x -> Complex x -> Maybe (ComplexMap (Complex ()) (Complex x))
-cpxMapVertex x c = case isVertex x c of
-  True  -> Just $ ComplexMap (cpxTerminal ()) c (OrdMap $ const x)
-  False -> Nothing
-
---------------------------------------------------------------------------------
-Just a = cpxBorder $ complex $ [set "abc"]
-Just b = cpxBorder $ complex $ [set [0::N .. 2]]
-c = cpxProduct a b
-
-p1 = ComplexMap c a (OrdMap fst)
-p2 = ComplexMap c b (OrdMap snd)
-
-u = complex [Set "abe"]
-v = complex [Set "abc",Set "bce",Set "ae"]
-
-f = ComplexMap u v (OrdMap id)
-
+f :: s ~ []
+  => Any n -> ComplexMap s n (Complex s n Char) (Complex s n Char) 
+f n = ComplexMap (singularComplex n a) (singularComplex n b) (OrdMap id)
 -}
+
+--------------------------------------------------------------------------------
+-- cpxsChainMapRep -
+
+cpxsChainMapRep :: ( Ring r, Commutative r, SimplicalTransformable s x y, Typeable s
+                   )
+  => ComplexMap s n (Complex s n x) (Complex s n y)
+  -> FinList (n+3) (Representable r (ChainHom r s) (Chain r s x) (Chain r s y))
+cpxsChainMapRep (ComplexMap csx csy f)
+  = amap1 (uncurry (rep f)) (amap1 snd (cpxxs csx) `F.zip` amap1 snd (cpxxs csy)) where
+
+  rep :: ( Ring r, Commutative r, SimplicalTransformable s x y, Typeable s)
+    => Map EntOrd x y -> Set (s x) -> Set (s y)
+    -> Representable r (ChainHom r s) (Chain r s x) (Chain r s y)
+  rep f ssx ssy = chainHomRep (ChainMap ssx ssy f)
+
+cpxsChainMapRepZ :: (r ~ Z, SimplicalTransformable s x y, Typeable s)
+  => ComplexMap s n (Complex s n x) (Complex s n y)
+  -> FinList (n+3) (Representable r (ChainHom r s) (Chain r s x) (Chain r s y))
+cpxsChainMapRepZ = cpxsChainMapRep
+
+
+a = setComplex [Set [(0,0),(0,1),(1,1)],Set [(0,0),(1,0),(1,1)]] :: SetComplex (N,N)
+b = setComplex [Set [0,1]] :: SetComplex N
+
+p1 = SetComplexMap a b (Map fst)
+p2 = SetComplexMap a b (Map snd)
+
+f :: s ~  [] => Any n -> ComplexMap s n (Complex s n (N,N)) (Complex s n N) 
+f n = ComplexMap (singularComplex n a) (singularComplex n b) (Map fst)
